pyproject.toml
[project]
name = "center-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "openai>=2.30.0",
    "paho-mqtt>=2.1.0",
    "python-multipart>=0.0.20",
    "python-socketio>=5.13.0",
    "requests>=2.33.1",
    "uvicorn>=0.34.2",
]
